requirements.txt
aiosqlite==0.13.0
astroid==2.4.2
asynctest==0.13.0
attrs==19.3.0
autopep8==1.5.3
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
fastapi==0.55.1
h11==0.9.0
httptools==0.1.1
idna==2.10
iso8601==0.1.12
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
passlib==1.7.2
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.5.1
PyJWT==1.7.1
pylint==2.5.3
pyparsing==2.4.7
PyPika==0.37.11
pytest==5.4.3
python-multipart==0.0.5
requests==2.24.0
six==1.15.0
starlette==0.13.2
toml==0.10.1
tortoise-orm==0.16.13
typing-extensions==3.7.4.2
urllib3==1.25.9
utils==1.0.1
uvicorn==0.11.5
uvloop==0.14.0
wcwidth==0.2.5
websockets==8.1
wrapt==1.12.1
